Update CSS
Update HTML and CSS to show the image as background image for the Home page.

Check footer issue-it's covering the whole website for some unknown reasons. Need to be solved if anyone can help.

Remove ul from the .nav-links CSS. It was selecting the ul inside the ul. It should be working now and should show a red line under the nav links.