update jquery for stone section
Add jquery so that everytime the user click on Read More, it expands the text.
This allow us to not write three functions to perform the same action.
If we need to write three functions we need to have different IDs assigned to each card.
Now I need to figure out how to update the word 'Read More' to 'Read less'